backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
weaviate-client==4.9.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
openai>=1.0.0
pdf2image>=1.16.0
Pillow>=10.0.0

